+++ src/onegov.ballot/onegov/ballot/models/election/list.py
@@ -58,8 +58,10 @@ class List(Base, TimestampMixin):
     #: a (proporz) list contains votes from other other lists
     panachage_results = relationship(
         'PanachageResult',
+        primaryjoin=(
+            'foreign(PanachageResult.target) == cast(List.id, TEXT)'
+        ),
         cascade='all, delete-orphan',
-        backref=backref('list'),
         lazy='dynamic'
     )