+++ src/onegov.ballot/onegov/ballot/models/election/panachage_result.py
@@ -2,7 +2,6 @@ from onegov.core.orm import Base
 from onegov.core.orm.mixins import TimestampMixin
 from onegov.core.orm.types import UUID
 from sqlalchemy import Column
-from sqlalchemy import ForeignKey
 from sqlalchemy import Integer
 from sqlalchemy import Text
 from uuid import uuid4
@@ -17,11 +16,14 @@ class PanachageResult(Base, TimestampMixin):
     #: identifies the result
     id = Column(UUID, primary_key=True, default=uuid4)
 
+    #: the owner of this result
+    owner = Column(Text, nullable=True)
+
     #: the target this result belongs to
-    target_list_id = Column(UUID, ForeignKey('lists.id'), nullable=False)
+    target = Column(Text, nullable=False)
 
     #: the source this result belongs to
-    source_list_id = Column(Text, nullable=False)
+    source = Column(Text, nullable=False)
 
     # votes
     votes = Column(Integer, nullable=False, default=lambda: 0)